Fix Kotlin lint warnings reported by downstream sync (closes #531) (#536)
## Summary

Addresses the seven categories of warnings caught by static analysis
during the CL 897461783 sync (#531). Mostly mechanical, with two
substantive bug fixes (now with regression tests), plus a CI knob to
keep these from coming back.

## Categories addressed

| # | Category | Sites | Fix |
|---|---|---|---|
| 1 | Ignored return values | 14 | `val unused = stub.X(...)` per the
issue's guidance, plus a real bug fix (see below) |
| 2 | Deprecated `Replica.egressPort` / `setEgressPort` | 13 (more than
the issue listed) | Read sites use the existing `replicaPort` helper
(now public, in `ArchitectureHelpers`); write sites use a new
`portToBytes` companion helper. Production sites in `stf/Runner.kt`
updated too |
| 3 | `forEach` → `for` loop | 4 | Mechanical |
| 4 | Unnecessary `!!` + redundant conversion | several | `checkNotNull`
to a local; dropped a `Long.toLong()` and a `UInt32.toInt()` |
| 5 | Suspend in `finally` (`P4RuntimeService.kt:667`) | 1 | Wrapped in
`withContext(NonCancellable)` |
| 6 | Property syntax (`P4RuntimeWriteErrorTest.kt`) | 2 |
`getCanonicalCode()` → `canonicalCode` |
| 7 | Unresolved + indirect-dep refs | 3 | Added explicit imports
(`com.google.rpc.Code`, `com.google.rpc.Status as RpcStatus`) |
| — | Unrelated `condition is always 'true'` in `TypeTranslator` | 1 |
Dropped the dead null check on a non-null return |

## Substantive bug fixes uncovered along the way (with regression tests)

- **`DataplaneService.subscribeResults` was silently dropping `trySend`
failures.** If the subscriber's channel was closed or full, the packet
result was lost without surfacing the error. Now closes the flow on send
failure, preserving the `SubscribeResults` contract.
- **Test:** `DataplaneServiceTest.SubscribeResults flow closes when
subscriber falls behind`. Verified: fails on the unfixed code (collector
hangs past the 10s timeout because dropped packets never close the
flow), passes on the fix.

- **`P4RuntimeService.streamChannel`'s `finally` block called a
cancellable suspend (`handleDisconnect`).** If the stream coroutine was
cancelled mid-disconnect, the controller would stay in the `controllers`
map and block role primary recomputation. Wrapped in
`withContext(NonCancellable)`.
- **Test:** `P4RuntimeConformanceTest.74b - abrupt cancel of primary
stream still releases its arbitration slot`. Verified: fails on the
unfixed code (new lower-election controller gets `ALREADY_EXISTS`
because the stale primary lingers in the map), passes on the fix. 8
iterations to widen the cancellation race window.

## CI: keep these from coming back

Set `kt_kotlinc_options(warn = "error")` on the Kotlin toolchain in
`bazel/BUILD.bazel`. This treats Kotlin compiler warnings as errors
locally and in CI, catching ~5 of the 7 categories in #531 (deprecated,
forEach style, !! redundancy, suspend-in-finally, redundant conversion)
at compile time. Strict-deps was already at `error`. Categories not
covered: `@CheckReturnValue` enforcement (would need ErrorProne or a
custom detekt rule) and the indirect-dep flavor that the downstream sees
but Bazel's `strict_kotlin_deps` doesn't.

## Performance

Hot-path files touched (`Interpreter.kt`, `TableStore.kt`,
`PSAArchitecture.kt`, `PNAArchitecture.kt`, `DefaultValues.kt`). Ran
`DataplaneBenchmark` 3 times before/after; no regression, modest wins on
small-state configs from the `forEach → for` conversion in
`TableStore.deepCopy` (no lambda allocation per snapshot).

| Config | Before (pps, 3-run mean) | After (pps) | Δ |
|---|---:|---:|---:|
| direct 100 | 3,644 | 3,878 | +6.4% |
| direct 1k | 3,515 | 3,729 | +6.1% |
| direct 10k | 2,516 | 2,512 | -0.2% |
| wcmp×16 | 1,696 | 1,721 | +1.5% |
| wcmp×16+mirr | 1,185 | 1,202 | +1.4% |

## Test plan

- [x] `bazel build //...`
- [x] `bazel test //... --test_tag_filters=-heavy` (64/64 + 2 new pass)
- [x] `./tools/format.sh`, `./tools/lint.sh` clean
- [x] `DataplaneBenchmark`: no regression
- [x] Verified `warn = "error"` rejects newly-introduced warnings
- [x] Verified both new tests fail on the unfixed code (regression
coverage proven)
- [ ] CI heavy tests

+++ simulator/DefaultValues.kt
@@ -35,9 +35,7 @@ internal fun defaultValue(type: Type, types: Map<String, TypeDecl>): Value =
       HeaderStackVal(
         elementTypeName = type.headerStack.elementType,
         headers =
-          MutableList(type.headerStack.size.toInt()) {
-            defaultValue(type.headerStack.elementType, types)
-          },
+          MutableList(type.headerStack.size) { defaultValue(type.headerStack.elementType, types) },
       )
     else -> UnitVal // varbit<N>: variable-length; no fixed default
   }
